Set up initial HTML structure
Built sections based off the figma wireframe for the project. Tweaked
some sections to look different based on Scrimba tutorial for
portfolio site and after looking at other portfolio websites online.

Set up basic structure of navigation area with logo, an intro area
with a text and image, an area to display skills and an about
me section, and finally a project section.

Still to come is a contact section with a form at the bottom of
the page.