Create styles for Introduction Section with CSS
Using a media query for screen sizes 600px and larger, CSS Grid
is used to position and size the image, title, and subtitle
in the Introduction Section.

Before the query a position of relative is applied to the section
element that holds all of the introduction elements. This is in
order to give elements within it an absolute positioning to
move them about to achieve a desired effect.