Add CSS Styling to finish layout of Introduction Section
The image in the introduction section is now positioned correctly
as well as the subtitle and title. The subtitle now has added
padding and calculated width to make it stretch out the full
width of the introduction section. It sits behind the image now
which has an z-index of 2 to keep it infront of the subtitle
content.

Toying with these positions will come later after the
other sections are also 90% done like this section is.