Create multi-column layout and animation effects for About Me Section
The About section now has hover effects that bring attention to the
hovered over article while dimming the others to give the user
better visual focus. A display of flex was used to achieve the
article layout within the services div.

Each has the same height and width regardless of content size. This
was done using a flex-basis.